refactor(STORE): ROOM CONTEXT - NAMESPACE RENAME
